Introduced Initial testing of GOAP AI
So I spent some time developing a few GOAP header only libraries, and using them set up an AIAgent node that will be able to develop it's on action plans.
The  headers are: 
GOAPcore - where the very basic elements of GOAP happen. The planner, MultitypeArray "WorldState" and base actions are here.
GOAPAgents - Agents and behaviours are essentially containers for what actions, goals and elements of worldstates.
GOAPAI - Interfaces for Agents and Actions to create a running statemachine to use behaviours to find it's own plans.

TODO:
Currently The AI is picking up food from the bush and taking to the crate. It needs to not perform actions instantly and also include animations.
Will then be able to add more/different actions and behaviours.

=Godot related=
The Node starter class is an effort in delaying process function overrides until the game is running, rather than in the editor. I would like to look more into how _process is called in engine so that IU don't need this jank fix.