Begin Adding Animations for actions
I was having a fair amount of issues with loading animations with models, as fbx is not officially supported by godot. I had issues converting some animations and models to gltf, so I settled on this one little guy and just created my own simple animations.

A new Node class was created, and it's most likely unecessary, but it simply adds an interface to get an animation tree from it's scene in order for other scenes to get a refence to it. This was so I could keep AIagents and their models as separate scenes.