Adding GOAP as Resources: Change To Behaviours, Actions and Goals
Initially, I was working on being able to edit animations for actions in an easier method, tweaking some values inside the inspector would be a nice method. But as I was learning more about the Godot engine to expose the variables. It began to dawn on me that I could and should use the interfaces Godot provides, so that I will be able to easily edit user defined parts of GOAP.

Behaviours, Actions and Goals all now have an class Inheriting from Resource. A resource in Godot is essentially a container that holds various amounts of data and most importantly can be saved to disk. In using these I can use the inspector to set all the variables I wish. For example, I can set an the array of action resources that a behaviours will use, or I can set the animation timings that an action will use.

The last thing I'll need to do now for this  resource method to work is make the AIAgent Example use these instead of the predefined classes. I'll need to create a whole lot of different test resources for this to work. I figured I would push these changes now, as there may be a lot more hurdles before it works completely.