Creating Resource classes for GOAP
While adding resources to using Godot's interfaces. I have come across numerous problems. Using a resource that holds the keys and values of base GOAPWorldstate values seems to be an ok approach to inserting dynamic values into behaviours. The values are grabbed from the resource and added into the behaviours as they are set. However, There is an ERROR somewhere with a corrupted data heap. I need to track this down and to do this I will create some unit tests for the GOAP headers to ensure that the problem does not exist there.